move depobs.scanner.clients to depobs.clients
