+++ depobs/worker/tasks.py
@@ -46,10 +46,10 @@ from depobs.database.models import (
 import depobs.worker.scoring as scoring
 import depobs.worker.validators as validators
 
-# import exc_to_str to resolve import cycle for the following depobs.scanner.clients
+# import exc_to_str to resolve import cycle for the following depobs.clients
 from depobs.util.traceback_util import exc_to_str as _
-from depobs.scanner.clients.npmsio import fetch_npmsio_scores, NPMSIOClientConfig
-from depobs.scanner.clients.npm_registry import (
+from depobs.clients.npmsio import fetch_npmsio_scores, NPMSIOClientConfig
+from depobs.clients.npm_registry import (
     fetch_npm_registry_metadata,
     NPMRegistryClientConfig,
 )
